fix: Moved Size and Position into CLI because xtas k does not depend on inlyne and removed unnecessary & in main.rs
