+++ src/opts/config.rs
@@ -1,9 +1,8 @@
 use std::fs::{create_dir_all, read_to_string};
 use std::io::Write;
 use std::path::{Path, PathBuf};
-use std::str::FromStr;
 
-use super::ThemeType;
+use super::{Position, Size, ThemeType};
 use crate::color;
 use crate::keybindings::Keybindings;
 
@@ -87,53 +86,6 @@ pub enum MetricsExporter {
 #[serde(default)]
 pub struct DebugSection {
     pub metrics: Option<MetricsExporter>,
-}
-
-#[derive(Deserialize, Clone, Debug, Default, PartialEq)]
-pub struct Position {
-    pub x: i32,
-    pub y: i32,
-}
-
-impl FromStr for Position {
-    type Err = &'static str;
-
-    fn from_str(input: &str) -> Result<Self, Self::Err> {
-        let parts: Vec<&str> = input.split(',').collect();
-        if parts.len() != 2 {
-            return Err("Invalid format for Position: expected format <x>,<y>");
-        }
-        let x = parts[0]
-            .parse::<i32>()
-            .map_err(|_| "Invalid x-coordinate: not a valid integer")?;
-        let y = parts[1]
-            .parse::<i32>()
-            .map_err(|_| "Invalid y-coordinate: not a valid integer")?;
-        Ok(Position { x, y })
-    }
-}
-
-#[derive(Deserialize, Clone, Debug, Default, PartialEq)]
-pub struct Size {
-    pub width: u32,
-    pub height: u32,
-}
-impl FromStr for Size {
-    type Err = &'static str;
-
-    fn from_str(input: &str) -> Result<Self, Self::Err> {
-        let parts: Vec<&str> = input.split('x').collect();
-        if parts.len() != 2 {
-            return Err("Invalid format for Size: expected format <width>x<height>");
-        }
-        let width = parts[0]
-            .parse::<u32>()
-            .map_err(|_| "Invalid width: not a valid integer")?;
-        let height = parts[1]
-            .parse::<u32>()
-            .map_err(|_| "Invalid height: not a valid integer")?;
-        Ok(Size { width, height })
-    }
 }
 
 #[derive(Deserialize, Debug, Default, PartialEq)]
